% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcAGH.R
\name{arcbin.fit}
\alias{arcbin.fit}
\title{Fitting functions for categorical data.}
\usage{
arcbin.fit(
  x,
  y,
  f1,
  f2,
  obj_glm,
  nq = 25,
  niter = 10,
  get_effects = FALSE,
  get_se = TRUE
)
}
\arguments{
\item{x}{Model matrix.}

\item{y}{Response vector (binary data).}

\item{f1}{Factor representing the row effects.}

\item{f2}{Factor representing the columns effects.}

\item{obj_glm}{Object representing the glm fit of a probit model
as returned by \code{glm}.}

\item{nq}{Number of quadrature points. Default is 25.}

\item{niter}{Number of Newton-Raphson iteration to locate
the mode of each random effect. Default is 10.}

\item{get_effects}{Returns the random effects prediction.
Default is FALSE.}

\item{get_se}{Returns the standard errors of beta coefficients.
Default is TRUE.}
}
\value{
A list  containing the following components:
\item{\code{beta}}{the vector of estimated coefficients.}
\item{\code{beta.se}}{the estimated standard errors of the
coefficients.}
\item{\code{sigmaA, sigmaB}}{the estimated standard deviation of
row and column random effects.}
\item{\code{a.est, b.est}}{the predicted random effects.}
}
\description{
Fit a binary probit model with crossed random effects by the ARC method.
}
\examples{
library(glmm)
data(salamander)
y <- salamander$Mate
mod.glm <- glm(Mate ~ Cross - 1, family = binomial("probit"), data = salamander)
x <- model.matrix(mod.glm)
fit.arc <- arcbin.fit(x, y, f1 = salamander$Male, f2 = salamander$Female,
                      obj_glm = mod.glm)

}
