% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcAGH.R
\name{arcord.fit}
\alias{arcord.fit}
\title{Fit an ordinal probit model with crossed random effects by the ARC method.}
\usage{
arcord.fit(
  x,
  y,
  f1,
  f2,
  obj_polr,
  nq = 25,
  niter = 10,
  get_effects = FALSE,
  get_se = TRUE
)
}
\arguments{
\item{x}{Model matrix.}

\item{y}{Response vector (ordinal data with $K$ categories).}

\item{f1}{Factor representing the row effects.}

\item{f2}{Factor representing the columns effects.}

\item{obj_polr}{Object representing the fit of an ordinal probit model
as returned by \code{MASS::polr}.}

\item{nq}{Number of quadrature points. Default is 25.}

\item{niter}{Number of Newton-Raphson iteration to locate
the mode of each random effect. Default is 10.}

\item{get_effects}{Returns the random effects prediction.
Default is FALSE.}

\item{get_se}{Returns the standard errors of model coefficients.
Default is TRUE.}
}
\value{
A list  containing the following components:
\item{\code{beta}}{the vector of estimated coefficients.}
\item{\code{beta.se}}{the estimated standard errors of the
coefficients.}
\item{\code{alpha}}{the vector of estimated thresholds.}
\item{\code{alpha.se}}{the estimated standard errors of the
thresholds.}
\item{\code{sigmaA, sigmaB}}{the estimated standard deviation of
row and column random effects.}
\item{\code{a.est, b.est}}{the predicted random effects.}
}
\description{
Fit an ordinal probit model with crossed random effects by the ARC method.
}
\examples{
library(lme4)
mod.polr <- MASS::polr(factor(y) ~ studage + lectage + service + dept, method = "probit",
                       data = InstEval)
x <- model.matrix(mod.polr)[,-1]
fit.arc <- arcord.fit(x, InstEval$y, f1 = InstEval$s, f2 = InstEval$d, obj_polr = mod.polr)

}
