% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LapProbit.R
\name{lapbin.fit}
\alias{lapbin.fit}
\title{Laplace-based functions for categorical data.}
\usage{
lapbin.fit(x, y, f1, f2, gamma, lsA = log(0.5), lsB = log(0.5))
}
\arguments{
\item{x}{Model matrix.}

\item{y}{Response vector (binary data).}

\item{f1}{Factor representing the row effects.}

\item{f2}{Factor representing the columns effects.}

\item{gamma}{Numerical vector of estimated coefficients of binary data probit model.}

\item{lsA}{Numerical, starting value for logarithm of row effects standard deviation.}

\item{lsB}{Numerical, starting value for logarithm of column effects standard deviation.}
}
\value{
A list  containing the following components:
\item{\code{beta}}{the vector of estimated coefficients.}
\item{\code{alpha}}{the vector of estimated thresholds.}
\item{\code{sigmaA, sigmaB}}{the estimated standard deviation of
row and column random effects.}
}
\description{
Approximate MLE of a binary data probit model with crossed random effects based on the
1st-order Laplace's approximation.
}
\examples{
library(glmm)
data(salamander)
y <- salamander$Mate
mod.glm <- glm(Mate ~ Cross - 1, family = binomial("probit"), data = salamander)
x <- model.matrix(mod.glm)
fit.lap <- lapbin.fit(x, y, f1 = salamander$Male, f2 = salamander$Female, gamma = mod.glm$coef)

}
