% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcProbit.R
\docType{_PACKAGE}
\name{arcProbit}
\alias{arcProbit}
\title{arcProbit: All-row-col fitting of categorical data models with crossed random effects}
\description{
The \code{arcProbit} package can be used to estimate the parameters of mixed models for binary
and ordinal data with crossed random effects. It implements the ARC estimation proposed in
https://arxiv.org/abs/2308.15681. A vignette is available with some illustrative examples.
}
\references{
Bellio, R., Ghosh, S., Owen, A. and Varin, C. (2023). Consistent and Scalable Composite Likelihood Estimation of Probit Models
with Crossed Random Effects. https://arxiv.org/abs/2308.15681
}
